requirements.txt
asgiref==3.8.1
black==24.3.0
click==8.1.7
Django==4.2.11
django-extensions==3.2.3
django-split-settings==1.3.0
djangorestframework==3.15.1
flake8==6.1.0
gunicorn==21.2.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
